Move WSGI.parse_request() to util()
